+++ src/main/java/de/markusbordihn/cats/commands/CatFollowCommand.java
@@ -25,7 +25,6 @@ import com.hypixel.hytale.server.core.Message;
 import com.hypixel.hytale.server.core.command.system.CommandContext;
 import com.hypixel.hytale.server.core.command.system.arguments.types.ArgTypes;
 import com.hypixel.hytale.server.core.command.system.arguments.types.EntityWrappedArg;
-import com.hypixel.hytale.server.core.command.system.basecommands.AbstractWorldCommand;
 import com.hypixel.hytale.server.core.universe.world.World;
 import com.hypixel.hytale.server.core.universe.world.storage.EntityStore;
 import com.hypixel.hytale.server.npc.entities.NPCEntity;
@@ -34,7 +33,7 @@ import de.markusbordihn.cats.component.CatStateComponent;
 import de.markusbordihn.cats.data.CatState;
 import javax.annotation.Nonnull;
 
-final class CatFollowCommand extends AbstractWorldCommand implements CatCommandHelper {
+final class CatFollowCommand extends CatCommand {
 
   @Nonnull private final EntityWrappedArg entityArg;